refactor: remove API_Filter, add R3 two-phase W entry+taint, fast-path skip J for no_input
- Remove API_Filter from engine/lean_engine pipelines
- R3-W split into Phase 1 (entry detection) and Phase 2 (taint analysis)
- Same session for both phases; Phase 2 only runs if has_external_input=true
- Fast path: has_external_input=false skips R3-J entirely
- Fix API_Filter priority truncation (int→float) for slot fairness
- Data-backed: 0/23 conflicts when W says no_external_input

+++ app/agent_slots.py
@@ -233,7 +233,7 @@ class AgentProcessSlotManager:
             self._sequence += 1
             ticket = AgentSlotTicket(
                 sequence=self._sequence,
-                priority=int(priority),
+                priority=float(priority),
                 task_id=str(task_id or "").strip() or None,
                 stage_key=str(stage_key or "").strip() or None,
                 role_kind=str(role_kind or "").strip() or None,
